+++ kubernetes-model-generator/openshift-model-miscellaneous/cmd/generate/generate.go
@@ -32,6 +32,7 @@ import (
   networkattachment "github.com/k8snetworkplumbingwg/network-attachment-definition-client/pkg/apis/k8s.cni.cncf.io/v1"
   baremetal "github.com/metal3-io/baremetal-operator/apis/metal3.io/v1alpha1"
   clusternetworkoperator "github.com/openshift/cluster-network-operator/pkg/apis/network/v1"
+  v1beta1_clusterapiprovidermetal3 "github.com/metal3-io/cluster-api-provider-metal3/api/v1beta1"
   networkoperator "github.com/openshift/api/networkoperator/v1"
   apiserver "github.com/openshift/api/apiserver/v1"
   imageregistry "github.com/openshift/api/imageregistry/v1"
@@ -64,6 +65,10 @@ type Schema struct {
   ConfigList                               imageregistry.ConfigList
   APIRequestCount                          apiserver.APIRequestCount
   APIRequestCountList                      apiserver.APIRequestCountList
+  Metal3Remediation                        v1beta1_clusterapiprovidermetal3.Metal3Remediation
+  Metal3RemediationList                    v1beta1_clusterapiprovidermetal3.Metal3RemediationList
+  Metal3RemediationTemplate                v1beta1_clusterapiprovidermetal3.Metal3RemediationTemplate
+  Metal3RemediationTemplateList            v1beta1_clusterapiprovidermetal3.Metal3RemediationTemplateList
 }
 
 func main() {
@@ -81,6 +86,7 @@ func main() {
     {"github.com/openshift/cloud-credential-operator/pkg/apis/cloudcredential/v1", "cloudcredential", "io.fabric8.openshift.api.model.miscellaneous.cloudcredential.v1", "os_cloudcredential_v1_", true},
     {"github.com/k8snetworkplumbingwg/network-attachment-definition-client/pkg/apis/k8s.cni.cncf.io/v1", "k8s.cni.cncf.io", "io.fabric8.openshift.api.model.miscellaneous.cncf.cni.v1", "os_cncf_cni_v1_", true},
     {"github.com/metal3-io/baremetal-operator/apis/metal3.io/v1alpha1", "metal3.io", "io.fabric8.openshift.api.model.miscellaneous.metal3.v1alpha1", "os_metal3_v1alpha1_", true},
+    {"github.com/metal3-io/cluster-api-provider-metal3/api/v1beta1", "infrastructure.cluster.x-k8s.io", "io.fabric8.openshift.api.model.miscellaneous.metal3.v1beta1", "os_metal3_v1beta1_", true},
     {"github.com/openshift/cluster-network-operator/pkg/apis/network/v1", "network.operator", "io.fabric8.openshift.api.model.miscellaneous.network.operator.v1", "os_network_operator_v1_", true},
     {"github.com/openshift/api/networkoperator/v1", "network.operator", "io.fabric8.openshift.api.model.miscellaneous.network.operator.v1", "os_network_operator_v1_", true},
     {"github.com/openshift/api/imageregistry/v1", "imageregistry.operator", "io.fabric8.openshift.api.model.miscellaneous.imageregistry.operator.v1", "os_imageregistry_v1_", true},